Pull parsing logic into distinct file and remove invalid continue
