+++ packages/safe-chain/src/packagemanager/rush/createRushPackageManager.js
@@ -1,5 +1,6 @@
 import { runRushCommand } from "./runRushCommand.js";
 import { resolvePackageVersion } from "../../api/npmApi.js";
+import { parsePackagesFromRushAddArgs } from "./parsing/parsePackagesFromRushAddArgs.js";
 
 /**
  * @returns {import("../currentPackageManager.js").PackageManager}
@@ -22,9 +23,7 @@ async function scanRushAddCommand(args) {
     return [];
   }
 
-  const parsedSpecs = extractRushAddPackageSpecs(args)
-    .map((spec) => parsePackageSpec(spec))
-    .filter((spec) => spec !== null);
+  const parsedSpecs = parsePackagesFromRushAddArgs(args.slice(1));
 
   const resolvedVersions = await Promise.all(
     parsedSpecs.map(async (parsed) => {
@@ -62,79 +61,4 @@ function getRushCommand(args) {
   }
 
   return args[0]?.toLowerCase();
-}
-
-/**
- * @param {string[]} args
- * @returns {string[]}
- */
-function extractRushAddPackageSpecs(args) {
-  const packageSpecs = [];
-
-  for (let i = 1; i < args.length; i++) {
-    const arg = args[i];
-    if (!arg) {
-      continue;
-    }
-
-    if (arg === "--package" || arg === "-p") {
-      const next = args[i + 1];
-      if (next && !next.startsWith("-")) {
-        packageSpecs.push(next);
-        i += 1;
-      }
-      continue;
-    }
-
-    if (arg.startsWith("--package=")) {
-      const value = arg.slice("--package=".length);
-      if (value) {
-        packageSpecs.push(value);
-      }
-      continue;
-    }
-
-    if (!arg.startsWith("-")) {
-      packageSpecs.push(arg);
-    }
-  }
-
-  return packageSpecs;
-}
-
-/**
- * @param {string} spec
- * @returns {{name: string, version: string | null} | null}
- */
-function parsePackageSpec(spec) {
-  const value = removeAlias(spec.trim());
-  if (!value) {
-    return null;
-  }
-
-  const lastAtIndex = value.lastIndexOf("@");
-  if (lastAtIndex > 0) {
-    return {
-      name: value.slice(0, lastAtIndex),
-      version: value.slice(lastAtIndex + 1),
-    };
-  }
-
-  return {
-    name: value,
-    version: null,
-  };
-}
-
-/**
- * @param {string} spec
- * @returns {string}
- */
-function removeAlias(spec) {
-  const aliasIndex = spec.indexOf("@npm:");
-  if (aliasIndex !== -1) {
-    return spec.slice(aliasIndex + 5);
-  }
-
-  return spec;
 }
